Refactoring - PN-222
- 임시저장
