+++ src/app/admin/page.tsx
@@ -2,7 +2,9 @@ import React from 'react';
 import {getServerSession} from "next-auth";
 import {options} from "@/app/api/auth/[...nextauth]/options";
 import AdminTopCard from "@/components/ui/card/AdminTopCard";
-import Sidebar from "@/components/Sidebar";
+import Sidebar from "@/components/admin/Sidebar";
+import BarChart from "@/components/ui/chart/BarChart";
+import RecentOrders from "@/components/admin/RecentOrders";
 
 const AdminPage = async () => {
     const session = await getServerSession(options);
@@ -19,6 +21,10 @@ const AdminPage = async () => {
             <main className="w-full ml-20 bg-gray-50">
                 <AdminTopCard data={data}/>
                 {/*<AdminHistoryItem/>*/}
+                <div className='p-4 grid md:grid-cols-3 grid-cols-1 gap-4'>
+                    <BarChart/>
+                    <RecentOrders/>
+                </div>
             </main>
         </div>
     );
