+++ src/components/admin/RecentOrders.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const RecentOrders = () => {
+    return (
+        <div>
+            RecentOrders
+        </div>
+    );
+};
+
+export default RecentOrders;
