+++ src/components/admin/Scanner.tsx
