+++ src/components/ui/chart/BarChart.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const BarChart = () => {
+    return (
+        <div>
+            BarChart
+        </div>
+    );
+};
+
+export default BarChart;
